added Object Class
added an object class. this reduces the memory overhead when lots of the same model is wanting to be drawn. the vertices only need to be held in memory once and then moved around to the correct position before it is drawn to the screen

Also added a objectArray class which makes it easier to run functions on all objects at the same time i.e draw()