Displaying Atoms
Now Displays the atoms of a chemical in a set colour based on the element.
Hydrogen is also set to be smaller than all other elements to make organic chemicals easier to see.
I also tidied up the chemical input window to call on another class for its functionality. This will make any changes to UI easier to implement.
currently bonds are not being displayed.
and it only displays the 2D structure. This will be changed in future development.