Fixed Render Texture Not Resizing
render textures for the 2D and 3D visualiser should now correctly resize when  the size changes.

I cannot update it when the window changes size directly as otherwise when resizing it creates lots of render textures (which are then unloaded) creating a large jump in memory usage. This still happens if the user repeatedly resizes the window however it is minimised.